completed processing - look at descr for format
title_word_counts is a dictionary. The keys are strings containing each
title we parse. The value is a separate dictionary that keeps track of
the word count for each title. The keys of this dictionary are strings
containing all of the words of every title we parse. The value for each
key is the number of times the word appears in the current title.